De-specify all examples so they will run on any MCU, assuming that the default pins are used and the PHY doesn't require a specific pin value to enable it.
