+++ examples/rtic-echo.rs
@@ -6,7 +6,7 @@
 //! Starts a TCP echo server on port `1337` at `ADDRESS`. `ADDRESS` is `10.0.0.1/24` by default.
 //!
 //! By default, the example assumes that the default RMII pins are used.
-//! To use it on an stm32-nucleo-f746zg dev board, the `rtic-echo-example-altpin` feature should be enabled. This may work on other
+//! To use it on an stm32-nucleo-f746zg dev board, the `nucleo-pins` feature should be enabled. This may work on other
 //! boards, but that hasn't been tested so your mileage may vary.
 //!
 //! To run this, install `probe-run` (`cargo install probe-run --version '~0.3'`), and ensure that `probe-run` can
@@ -22,19 +22,19 @@ use smoltcp::{
     wire::{self, IpAddress, Ipv4Address},
 };
 
+mod common;
+
 const ADDRESS: (IpAddress, u16) = (IpAddress::Ipv4(Ipv4Address::new(10, 0, 0, 1)), 1337);
+const MAC: [u8; 6] = [0x00, 0x01, 0x02, 0x03, 0x04, 0x05];
 
 #[rtic::app(device = stm32_eth::stm32, dispatchers = [SPI1])]
 mod app {
 
+    use crate::common::EthernetPhy;
+
     use systick_monotonic::Systick;
 
-    use stm32_eth::{
-        hal::{gpio::GpioExt, rcc::RccExt},
-        EthernetDMA, RxRingEntry, TxRingEntry,
-    };
-
-    use fugit::RateExtU32;
+    use stm32_eth::{EthernetDMA, RxRingEntry, TxRingEntry};
 
     use smoltcp::{
         iface::{self, Interface, SocketHandle},
@@ -46,15 +46,13 @@ mod app {
     use crate::NetworkStorage;
 
     #[local]
-    struct Local {}
-
-    #[shared]
-    struct Shared {
-        #[lock_free]
+    struct Local {
         interface: Interface<'static, &'static mut EthernetDMA<'static, 'static>>,
-        #[lock_free]
         tcp_handle: SocketHandle,
     }
+
+    #[shared]
+    struct Shared {}
 
     #[monotonic(binds = SysTick, default = true)]
     type Monotonic = Systick<1000>;
@@ -78,51 +76,18 @@ mod app {
         let rx_ring = cx.local.rx_ring;
         let tx_ring = cx.local.tx_ring;
 
-        #[cfg(any(feature = "stm32f7xx-hal", feature = "stm32f4xx-hal"))]
-        let (clocks, mono) = {
-            let rcc = p.RCC.constrain();
-
-            let clocks = rcc.cfgr.sysclk(100.MHz()).hclk(100.MHz()).freeze();
-
-            let mono = Systick::new(core.SYST, clocks.hclk().raw());
-
-            (clocks, mono)
-        };
-
-        #[cfg(feature = "stm32f1xx-hal")]
-        let (clocks, mono) = {
-            use stm32_eth::hal::flash::FlashExt;
-
-            let rcc = p.RCC.constrain();
-            let mut flash = p.FLASH.constrain();
-
-            let clocks = rcc
-                .cfgr
-                .sysclk(32.MHz())
-                .hclk(32.MHz())
-                .freeze(&mut flash.acr);
-
-            let mono = Systick::new(core.SYST, clocks.hclk().raw());
-
-            (clocks, mono)
-        };
-
-        #[cfg(feature = "stm32f1xx-hal")]
-        let gpiog = ();
-
-        #[cfg(any(feature = "stm32f4xx-hal", feature = "stm32f7xx-hal"))]
-        let gpiog = p.GPIOG.split();
+        let (clocks, gpio, ethernet) = crate::common::setup_clocks(p);
+        let mono = Systick::new(core.SYST, clocks.hclk().raw());
 
         defmt::info!("Setting up pins");
-        let (pins, mdio, mdc) =
-            crate::pins::get_pins(p.GPIOA.split(), p.GPIOB.split(), p.GPIOC.split(), gpiog);
+        let (pins, mdio, mdc) = crate::common::setup_pins(gpio);
 
         defmt::info!("Configuring ethernet");
 
         let (dma, mac) = stm32_eth::new_with_mii(
-            p.ETHERNET_MAC,
-            p.ETHERNET_MMC,
-            p.ETHERNET_DMA,
+            ethernet.mac,
+            ethernet.mmc,
+            ethernet.dma,
             rx_ring,
             tx_ring,
             clocks,
@@ -147,15 +112,13 @@ mod app {
 
         let neighbor_cache = smoltcp::iface::NeighborCache::new(&mut store.neighbor_cache[..]);
 
-        let mac_addr = [0x00, 0x01, 0x02, 0x03, 0x04, 0x05];
-
         let rx_buffer = TcpSocketBuffer::new(&mut store.tcp_socket_storage.rx_storage[..]);
         let tx_buffer = TcpSocketBuffer::new(&mut store.tcp_socket_storage.tx_storage[..]);
 
         let socket = TcpSocket::new(rx_buffer, tx_buffer);
 
         let mut interface = iface::InterfaceBuilder::new(dma, &mut store.sockets[..])
-            .hardware_addr(EthernetAddress::from_bytes(&mac_addr).into())
+            .hardware_addr(EthernetAddress::from_bytes(&crate::MAC).into())
             .neighbor_cache(neighbor_cache)
             .ip_addrs(&mut store.ip_addrs[..])
             .routes(routes)
@@ -168,7 +131,7 @@ mod app {
 
         interface.poll(now_fn()).unwrap();
 
-        if let Ok(mut phy) = crate::EthernetPhy::from_miim(mac, 0) {
+        if let Ok(mut phy) = EthernetPhy::from_miim(mac, 0) {
             defmt::info!(
                 "Resetting PHY as an extra step. Type: {}",
                 phy.ident_string()
@@ -179,22 +142,21 @@ mod app {
             defmt::info!("Not resetting unsupported PHY.");
         }
 
-        defmt::info!("Setup done.");
+        defmt::info!("Setup done. Listening at {}", crate::ADDRESS);
 
         (
-            Shared {
+            Shared {},
+            Local {
                 interface,
                 tcp_handle,
             },
-            Local {},
             init::Monotonics(mono),
         )
     }
 
-    #[task(binds = ETH, shared = [interface, tcp_handle], local = [data: [u8; 512] = [0u8; 512]], priority = 2)]
+    #[task(binds = ETH, local = [interface, tcp_handle, data: [u8; 512] = [0u8; 512]], priority = 2)]
     fn eth_interrupt(cx: eth_interrupt::Context) {
-        let (iface, tcp_handle, buffer) =
-            (cx.shared.interface, cx.shared.tcp_handle, cx.local.data);
+        let (iface, tcp_handle, buffer) = (cx.local.interface, cx.local.tcp_handle, cx.local.data);
 
         let interrupt_reason = iface.device_mut().interrupt_handler();
         defmt::debug!("Got an ethernet interrupt! Reason: {}", interrupt_reason);
@@ -255,225 +217,6 @@ impl TcpSocketStorage {
         Self {
             rx_storage: [0; 512],
             tx_storage: [0; 512],
-        }
-    }
-}
-
-mod pins {
-    pub use pins::*;
-
-    #[cfg(any(feature = "stm32f4xx-hal", feature = "stm32f7xx-hal",))]
-    mod pins {
-        use stm32_eth::{hal::gpio::*, EthPins};
-
-        pub type RefClk = PA1<Input>;
-        pub type Crs = PA7<Input>;
-        pub type TxD1 = PB13<Input>;
-        pub type RxD0 = PC4<Input>;
-        pub type RxD1 = PC5<Input>;
-
-        #[cfg(not(feature = "rtic-echo-example-altpin"))]
-        pub type TxEn = PB11<Input>;
-        #[cfg(not(feature = "rtic-echo-example-altpin"))]
-        pub type TxD0 = PB12<Input>;
-
-        #[cfg(all(feature = "rtic-echo-example-altpin"))]
-        pub type TxEn = PG11<Input>;
-        #[cfg(feature = "rtic-echo-example-altpin")]
-        pub type TxD0 = PG13<Input>;
-
-        pub type Mdio = PA2<Alternate<11>>;
-        pub type Mdc = PC1<Alternate<11>>;
-
-        pub fn get_pins(
-            gpioa: gpioa::Parts,
-            gpiob: gpiob::Parts,
-            gpioc: gpioc::Parts,
-            #[allow(unused_variables)] gpiog: gpiog::Parts,
-        ) -> (
-            EthPins<RefClk, Crs, TxEn, TxD0, TxD1, RxD0, RxD1>,
-            Mdio,
-            Mdc,
-        ) {
-            let ref_clk = gpioa.pa1.into_floating_input();
-            let crs = gpioa.pa7.into_floating_input();
-            let tx_d1 = gpiob.pb13.into_floating_input();
-            let rx_d0 = gpioc.pc4.into_floating_input();
-            let rx_d1 = gpioc.pc5.into_floating_input();
-
-            #[cfg(not(feature = "rtic-echo-example-altpin"))]
-            let (tx_en, tx_d0) = (
-                gpiob.pb11.into_floating_input(),
-                gpiob.pb12.into_floating_input(),
-            );
-
-            #[cfg(feature = "rtic-echo-example-altpin")]
-            let (tx_en, tx_d0) = (
-                gpiog.pg11.into_floating_input(),
-                gpiog.pg13.into_floating_input(),
-            );
-
-            #[cfg(feature = "stm32f4xx-hal")]
-            let (mdio, mdc) = {
-                let mut mdio = gpioa.pa2.into_alternate();
-                mdio.set_speed(Speed::VeryHigh);
-                let mut mdc = gpioc.pc1.into_alternate();
-                mdc.set_speed(Speed::VeryHigh);
-                (mdio, mdc)
-            };
-
-            #[cfg(any(feature = "stm32f7xx-hal"))]
-            let (mdio, mdc) = (
-                gpioa.pa2.into_alternate().set_speed(Speed::VeryHigh),
-                gpioc.pc1.into_alternate().set_speed(Speed::VeryHigh),
-            );
-
-            (
-                EthPins {
-                    ref_clk,
-                    crs,
-                    tx_en,
-                    tx_d0,
-                    tx_d1,
-                    rx_d0,
-                    rx_d1,
-                },
-                mdio,
-                mdc,
-            )
-        }
-    }
-
-    #[cfg(any(feature = "stm32f1xx-hal"))]
-    mod pins {
-        use stm32_eth::{
-            hal::gpio::{Alternate, Input, PushPull, *},
-            EthPins,
-        };
-
-        pub type RefClk = PA1<Input<Floating>>;
-        pub type Crs = PA7<Input<Floating>>;
-        pub type TxEn = PB11<Alternate<PushPull>>;
-        pub type TxD0 = PB12<Alternate<PushPull>>;
-        pub type TxD1 = PB13<Alternate<PushPull>>;
-        pub type RxD0 = PC4<Input<Floating>>;
-        pub type RxD1 = PC5<Input<Floating>>;
-
-        pub type Mdio = PA2<Alternate<PushPull>>;
-        pub type Mdc = PC1<Alternate<PushPull>>;
-
-        pub fn get_pins(
-            mut gpioa: gpioa::Parts,
-            mut gpiob: gpiob::Parts,
-            mut gpioc: gpioc::Parts,
-            _: (),
-        ) -> (
-            EthPins<RefClk, Crs, TxEn, TxD0, TxD1, RxD0, RxD1>,
-            Mdio,
-            Mdc,
-        ) {
-            let ref_clk = gpioa.pa1.into_floating_input(&mut gpioa.crl);
-            let mdio = gpioa.pa2.into_alternate_push_pull(&mut gpioa.crl);
-            let crs = gpioa.pa7.into_floating_input(&mut gpioa.crl);
-
-            let mdc = gpioc.pc1.into_alternate_push_pull(&mut gpioc.crl);
-            let rx_d0 = gpioc.pc4.into_floating_input(&mut gpioc.crl);
-            let rx_d1 = gpioc.pc5.into_floating_input(&mut gpioc.crl);
-
-            let tx_en = gpiob.pb11.into_alternate_push_pull(&mut gpiob.crh);
-            let tx_d0 = gpiob.pb12.into_alternate_push_pull(&mut gpiob.crh);
-            let tx_d1 = gpiob.pb13.into_alternate_push_pull(&mut gpiob.crh);
-
-            let pins = EthPins {
-                ref_clk,
-                crs,
-                tx_en,
-                tx_d0,
-                tx_d1,
-                rx_d0,
-                rx_d1,
-            };
-
-            (pins, mdio, mdc)
-        }
-    }
-}
-
-use ieee802_3_miim::{
-    phy::{
-        lan87xxa::{LAN8720A, LAN8742A},
-        BarePhy, KSZ8081R,
-    },
-    Miim, Pause, Phy,
-};
-
-/// An ethernet PHY
-pub enum EthernetPhy<M: Miim> {
-    /// LAN8720A
-    LAN8720A(LAN8720A<M>),
-    /// LAN8742A
-    LAN8742A(LAN8742A<M>),
-    /// KSZ8081R
-    KSZ8081R(KSZ8081R<M>),
-}
-
-impl<M: Miim> Phy<M> for EthernetPhy<M> {
-    fn best_supported_advertisement(&self) -> ieee802_3_miim::AutoNegotiationAdvertisement {
-        unimplemented!()
-    }
-
-    fn get_miim(&mut self) -> &mut M {
-        match self {
-            EthernetPhy::LAN8720A(phy) => phy.get_miim(),
-            EthernetPhy::LAN8742A(phy) => phy.get_miim(),
-            EthernetPhy::KSZ8081R(phy) => phy.get_miim(),
-        }
-    }
-
-    fn get_phy_addr(&self) -> u8 {
-        match self {
-            EthernetPhy::LAN8720A(phy) => phy.get_phy_addr(),
-            EthernetPhy::LAN8742A(phy) => phy.get_phy_addr(),
-            EthernetPhy::KSZ8081R(phy) => phy.get_phy_addr(),
-        }
-    }
-}
-
-impl<M: Miim> EthernetPhy<M> {
-    /// Attempt to create one of the known PHYs from the given
-    /// MIIM.
-    ///
-    /// Returns an error if the PHY does not support the extended register
-    /// set, or if the PHY's identifier does not correspond to a known PHY.
-    pub fn from_miim(miim: M, phy_addr: u8) -> Result<Self, ()> {
-        let mut bare = BarePhy::new(miim, phy_addr, Pause::NoPause);
-        let phy_ident = bare.phy_ident().ok_or(())?;
-        let miim = bare.release();
-        match phy_ident & 0xFFFFFFF0 {
-            0x0007C0F0 => Ok(Self::LAN8720A(LAN8720A::new(miim, phy_addr))),
-            0x0007C130 => Ok(Self::LAN8742A(LAN8742A::new(miim, phy_addr))),
-            0x00221560 => Ok(Self::KSZ8081R(KSZ8081R::new(miim, phy_addr))),
-            _ => Err(()),
-        }
-    }
-
-    /// Get a string describing the type of PHY
-    pub const fn ident_string(&self) -> &'static str {
-        match self {
-            EthernetPhy::LAN8720A(_) => "LAN8720A",
-            EthernetPhy::LAN8742A(_) => "LAN8742A",
-            EthernetPhy::KSZ8081R(_) => "KSZ8081R",
-        }
-    }
-
-    /// Initialize the PHY
-    pub fn phy_init(&mut self) {
-        match self {
-            EthernetPhy::LAN8720A(phy) => phy.phy_init(),
-            EthernetPhy::LAN8742A(phy) => phy.phy_init(),
-            EthernetPhy::KSZ8081R(phy) => {
-                phy.set_autonegotiation_advertisement(phy.best_supported_advertisement());
-            }
         }
     }
 }
